Lighting eveything!!!
Changelog:

Mathematics

- fixing LookAt function
- new perspective function (with fovy parameter)

Core

- fixing profiling in main loop

Shaders

- new shaders for primitive rendering, lighting and shading models, multiple
types of source of light, fog simulation, non-local viewer

Mesh

- new vertex attribute (normal) and dynamic loading
- Removing cube function
- new rendering methods (array, elements, instancing)
- batch rendering for primitive data

Texture

- new load method for reading pixels data from memory (not from file)

Camera

- refactoring camera class
- moving viewing info class and render context from rendering system to
camera

Lighting

- new classes and structs for lighting, material and shading info

Profile

- refactoring profile class

RAW Format File

- ppdating reading OBJ function to support more vertex attributes

Rendering System

- new method to clear OpenGL Framebufer (color and depth)

+++ core/src/keeventhandler.cpp
@@ -27,9 +27,12 @@
 #include <kecore.h>
 #include <keaux.h>
 
-// ----------------------------------------------------------------------------
-//  kengine::scene - member class definition
-// ----------------------------------------------------------------------------
+
+/*
+* 
+*  kengine::scene - member class definition
+* 
+*/
 kengine::scene::scene(core* engine, scene_manager* sceneManager)
 	: engineHandle{ engine }, sceneManagerHandle{ sceneManager }
 {
@@ -137,9 +140,11 @@ void kengine::scene::moveWindowEvent(int x, int y)
 }
 
 
-// ----------------------------------------------------------------------------
-//  kengine::scene_manager - member class definition
-// ----------------------------------------------------------------------------
+/*
+* 
+*  kengine::scene_manager - member class definition
+* 
+*/
 kengine::scene_manager::scene_manager()
 	: scenes{}
 {
@@ -167,7 +172,7 @@ void kengine::scene_manager::removeScene(size_t pos)
 
 	kengine::scene* s = scenes[pos];
 	//delete s;
-	scenes.erase(scenes.begin() + pos);
+	scenes.erase(scenes.begin() + static_cast<int>(pos));
 	delete s;
 
 }
